Almost there..

+++ src/main/java/Servlet/adminDeleteProduct.java
@@ -1,5 +1,6 @@
 package Servlet;
 
+import Beans.AccountSession;
 import DataAccess.ProductDao;
 
 import javax.servlet.*;
@@ -12,11 +13,19 @@ import java.sql.SQLException;
 public class adminDeleteProduct extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		HttpSession session = request.getSession(false);
+		AccountSession accountSession = (AccountSession) session.getAttribute("accountSession");
+		if (accountSession == null || !accountSession.isUserIsAdmin()) {
+			response.sendError(HttpServletResponse.SC_FORBIDDEN, "You're not authorized");
+			return;
+		}
+
 		ProductDao productDao = new ProductDao();
 		int userID = Integer.parseInt(request.getParameter("productid"));
 		try {
-			if(productDao.deleteProductByID(userID)){
-				response.sendRedirect(request.getContextPath()+"/adminDashboard");			}
+			if (productDao.deleteProductByID(userID)) {
+				response.sendRedirect(request.getContextPath() + "/adminDashboard");
+			}
 		} catch (SQLException throwables) {
 			throwables.printStackTrace();
 		}
