+++ src/main/java/Servlet/adminDeleteUser.java
@@ -1,5 +1,6 @@
 package Servlet;
 
+import Beans.AccountSession;
 import DataAccess.UserDao;
 
 import javax.servlet.*;
@@ -12,11 +13,18 @@ import java.sql.SQLException;
 public class adminDeleteUser extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		HttpSession session = request.getSession(false);
+		AccountSession accountSession = (AccountSession) session.getAttribute("accountSession");
+		if (accountSession == null || !accountSession.isUserIsAdmin()) {
+			response.sendError(HttpServletResponse.SC_FORBIDDEN, "You're not authorized");
+			return;
+		}
+
 		UserDao userDao = new UserDao();
 		int userID = Integer.parseInt(request.getParameter("userid"));
 		try {
-			if(userDao.deleteUserByID(userID)){
-				response.sendRedirect(request.getContextPath()+"/adminDashboard");
+			if (userDao.deleteUserByID(userID)) {
+				response.sendRedirect(request.getContextPath() + "/adminDashboard");
 			}
 		} catch (SQLException throwables) {
 			throwables.printStackTrace();
